pyproject.toml
[project]
name = "volt-be"
version = "0.1.0"
description = "Backend code for volt"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.0",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "ruff>=0.9.1",
]
